add trainmanager header file
