+++ src/rollingstock/trainmanager.cpp
@@ -4,6 +4,7 @@
 #include "bahnhof/track/track.h"
 #include "bahnhof/common/gamestate.h"
 #include "bahnhof/rollingstock/rollingstock.h"
+#include "bahnhof/rollingstock/trainmanager.h"
 #include "bahnhof/rollingstock/train.h"
 #include "bahnhof/routing/routing.h"
