+++ src/ui/tableline.cpp
@@ -4,6 +4,7 @@
 #include "bahnhof/graphics/rendering.h"
 #include "bahnhof/common/gamestate.h"
 #include "bahnhof/routing/routing.h"
+#include "bahnhof/rollingstock/train.h"
 
 namespace UI{
